Revert "Added Verner Solver"
This reverts commit 800f61d79f69750bed7489a70c617c8fbe58be81.
