--- lib/OrdinaryDiffEqVerner/Project.toml
@@ -1,19 +1,0 @@
-name = "OrdinaryDiffEqVerner"
-uuid = "563d1cf2-e57a-4bc9-9127-4017e9cead42"
-authors = ["ParamThakkar123 <paramthakkar864@gmail.com>"]
-version = "1.0.0"
-
-[deps]
-OrdinaryDiffEq = "1dea7af3-3e70-54e6-95c3-0bf5283fa5ed"
-
-[extras]
-DiffEqDevTools = "f3b72e0c-5b89-59e1-b016-84e28bfd966d"
-Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
-SafeTestsets = "1bc83da4-3b8d-516f-aca4-4fe02f6d838f"
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-
-[compat]
-julia = "1.10"
-
-[targets]
-test = ["DiffEqDevTools", "Random", "SafeTestsets", "Test"]
