--- lib/OrdinaryDiffEqVerner/src/alg_utils.jl
@@ -1,13 +1,0 @@
-isfsal(alg::Vern7) = false
-isfsal(alg::Vern8) = false
-isfsal(alg::Vern9) = false
-
-alg_order(alg::Vern6) = 6
-alg_order(alg::Vern7) = 7
-alg_order(alg::Vern8) = 8
-alg_order(alg::Vern9) = 9
-
-alg_stability_size(alg::Vern6) = 4.8553
-alg_stability_size(alg::Vern7) = 4.6400
-alg_stability_size(alg::Vern8) = 5.8641
-alg_stability_size(alg::Vern9) = 4.4762
