+++ src/OrdinaryDiffEq.jl
@@ -148,6 +148,7 @@ include("composite_algs.jl")
 include("caches/basic_caches.jl")
 include("caches/low_order_rk_caches.jl")
 include("caches/high_order_rk_caches.jl")
+include("caches/verner_caches.jl")
 include("caches/sdirk_caches.jl")
 include("caches/firk_caches.jl")
 include("caches/kencarp_kvaerno_caches.jl")
@@ -164,6 +165,7 @@ include("caches/qprk_caches.jl")
 
 include("tableaus/low_order_rk_tableaus.jl")
 include("tableaus/high_order_rk_tableaus.jl")
+include("tableaus/verner_tableaus.jl")
 include("tableaus/rosenbrock_tableaus.jl")
 include("tableaus/sdirk_tableaus.jl")
 include("tableaus/firk_tableaus.jl")
@@ -184,6 +186,7 @@ include("perform_step/exponential_rk_perform_step.jl")
 include("perform_step/explicit_rk_perform_step.jl")
 include("perform_step/low_order_rk_perform_step.jl")
 include("perform_step/high_order_rk_perform_step.jl")
+include("perform_step/verner_rk_perform_step.jl")
 include("perform_step/sdirk_perform_step.jl")
 include("perform_step/kencarp_kvaerno_perform_step.jl")
 include("perform_step/firk_perform_step.jl")
@@ -202,6 +205,7 @@ include("dense/interpolants.jl")
 include("dense/rosenbrock_interpolants.jl")
 include("dense/stiff_addsteps.jl")
 include("dense/low_order_rk_addsteps.jl")
+include("dense/verner_addsteps.jl")
 include("dense/high_order_rk_addsteps.jl")
 
 include("derivative_utils.jl")
