+++ src/alg_utils.jl
@@ -40,6 +40,9 @@ isfsal(alg::Rodas4) = false
 isfsal(alg::Rodas42) = false
 isfsal(alg::Rodas4P) = false
 isfsal(alg::Rodas4P2) = false
+isfsal(alg::Vern7) = false
+isfsal(alg::Vern8) = false
+isfsal(alg::Vern9) = false
 # Pseudo Non-FSAL
 isfsal(alg::PDIRK44) = false
 isfsal(alg::DImplicitEuler) = false
@@ -437,6 +440,10 @@ alg_order(alg::OwrenZen5) = 5
 alg_order(alg::DP5) = 5
 alg_order(alg::Tsit5) = 5
 alg_order(alg::DP8) = 8
+alg_order(alg::Vern6) = 6
+alg_order(alg::Vern7) = 7
+alg_order(alg::Vern8) = 8
+alg_order(alg::Vern9) = 9
 alg_order(alg::TanYam7) = 7
 alg_order(alg::TsitPap8) = 8
 alg_order(alg::RadauIIA3) = 3
@@ -675,6 +682,10 @@ ssp_coefficient(alg::SSPSDIRK2) = 4
 alg_stability_size(alg::ExplicitRK) = alg.tableau.stability_size
 alg_stability_size(alg::DP5) = 3.3066
 alg_stability_size(alg::Tsit5) = 3.5068
+alg_stability_size(alg::Vern6) = 4.8553
+alg_stability_size(alg::Vern7) = 4.6400
+alg_stability_size(alg::Vern8) = 5.8641
+alg_stability_size(alg::Vern9) = 4.4762
 
 alg_can_repeat_jac(alg::Union{OrdinaryDiffEqAlgorithm, DAEAlgorithm}) = false
 alg_can_repeat_jac(alg::OrdinaryDiffEqNewtonAdaptiveAlgorithm) = true
